Directly specify project to pack
Packages without specifying a project uses the sln file which has been
fine in most projects using the generalized workflow file.
With native projects, trying to pack the solution fails due to the
native project library, so optionally pass the specific project  to pack
and use a workflow that selects this specific project by parameter if
provided.