Fix for color map reference issue
AP Suite has a bug where internal references to color maps can be
collected if not attached to an appropriate IRenderData instance.
So handle pending a fix in a future patch release, do the following:

- Always attach new color maps to a render data instance when created
- Serialize and store the color map on invalidation, as the render data
  instances are removed, so the color map could be lost.
- Clear the color map on invalidation
- Rebuild the color map from the serialized store on recreation of the
  component render data.

This approach can be simplified back to single color map reference once
an AP Suite patch including the fix is released.