Native C++ PCA update
Move to a c++/cli wrapper project instead of P/Invoke.
Create some managed wrappers that can hold on to persisted native data
structures such as the underlying compressed Eigen::MatrixXf so it does
not have to repeatedly be recomputed. Access data through an analysis
object that accepts the grid parameters and matrix to provide
eigenvalues and component data. To avoid exposing the Eigen dependency,
wrap the matrix access behind a PIMPL pattern.

Profile and optimize to handle as large of matrices as possible. Limits
are reached at some point, as the LAPACK demands resolving to a
dense matrix, so at some point the non-empty voxels * number of features
still becomes excessivly large. Still attempt to support and profile
voxel size and bin size on the order of 0.1s and ensure it runs in
single digit minutes even on datasets of 10s of millions of ions.

It  appears that performance is more limited by the LA solvers which
seem to scale off the product of voxels and features.

Add a generalizable feature resolver system to organize switches between
different grid types without switch statements everywhere.

- Update to use new create series extension methods
- Ensure noise eigenvalue line is rendered on top